+++ plugins/util-config.js
@@ -1,8 +1,14 @@
 let classReg = /\bclass\s*=\s*"[^"]+/;
+let compound = defaultTag => {
+    return i => {
+        let a = i.seprateAttrs(defaultTag);
+        return `<${a.tag} ${a.attrs}>${i.content}</${a.tag}><div mx-view="${i.mxView}" ${a.viewAttrs} class="pa none"></div>`;
+    };
+};
 module.exports = {
     loaderType: 'cmd', //加载器类型
-    tmplFolder: 'tmpl', //模板文件夹，该文件夹下的js无法直接运行
-    srcFolder: 'src', //经该工具编译到的源码文件夹，该文件夹下的js可以直接运行
+    commonFolder: 'tmpl', //模板文件夹，该文件夹下的js无法直接运行
+    compiledFolder: 'src', //经该工具编译到的源码文件夹，该文件夹下的js可以直接运行
     cssnano: { //css压缩选项
         safe: true,
         autoprefixer: false
@@ -32,7 +38,7 @@ module.exports = {
         cssUrl: true, //样式中的url
         jsLoop: true, //js循环
         jsService: true, //js接口服务
-        jsThis: true, //js this别名
+        //jsThis: true, //js this别名
         tmplCmdSyntax: true,//命令语法检查
         tmplAttrImg: true, //模板img属性
         tmplDisallowedTag: true, //不允许的标签
@@ -48,17 +54,15 @@ module.exports = {
     tmplFileExtNames: ['html', 'haml', 'pug', 'jade', 'tpl'], //模板后缀
     tmplConstVars: {}, //模板中不会变的变量，减少子模板的分析
     tmplGlobalVars: {}, //模板中全局变量
-    //tmplUncheckTags: {}, //不检测的标签
-    //tmplAddEventPrefix: true, //mx事件增加前缀
     tmplAddViewsToDependencies: false, //是否把模板中的view做为依赖提前加载
     tmplOutputWithEvents: false, //输出事件
     tmplCompressVariable: true, //是否压缩模板中的变量
     tmplMultiBindEvents: false, //是否支持多个绑定
-    //tmplMESafeguard: false,//对象表达式保护
     tmplArtEngine: true,//类mustach模板引擎，因代码多参考artTempalte，因此以art命名
     tmplBindEvents: ['change'], //绑定表达式<%:expr%>绑定的事件
     tmplBindName: '@{sync.value.from.ui}', //绑定表达式<%:expr%>绑定的处理名称
     disableMagixUpdater: false,
+    magixUpdaterIncreasement: false,
     globalCss: [], //全局样式
     scopedCss: [], //全局但做为scoped使用的样式
     uncheckGlobalCss: [], //对某些全局样式不做检查
@@ -68,23 +72,21 @@ module.exports = {
     galleries: {
         mxRoot: 'app/gallery/',
         mxMap: {
-            'mx-suggest': {
-                tag: 'input'
+            'mx-suggest': compound('input'),
+            'mx-calendar.datepicker': compound('input'),
+            'mx-calendar.rangepicker': compound('input'),
+            'mx-color.picker': compound('input'),
+            'mx-suggest.index': compound('input'),
+            'mx-time.picker': compound('input'),
+            'mx-popover': compound('span'),
+            'mx-popover.index': compound('span'),
+            'mx-popconfirm': compound('a'),
+            'mx-popconfirm.index': compound('a'),
+            'mx-number': {
+                _class: 'input pr'
             },
-            'mx-calendar.datepicker': {
-                tag: 'input'
-            },
-            'mx-calendar.rangepicker': {
-                tag: 'input'
-            },
-            'mx-color.picker': {
-                tag: 'input'
-            },
-            'mx-suggest.index': {
-                tag: 'input'
-            },
-            'mx-time.picker': {
-                tag: 'input'
+            'mx-number.index': {
+                _class: 'input pr'
             },
             'mx-loading'() {
                 return `<div class="loading">
@@ -107,13 +109,13 @@ module.exports = {
     customTagProcessor() {
         return '';
     },
-    tmplPadCallArguments(name) { //模板中某些函数的调用，我们可以动态添加一些参数。
+    tmplPadCallArguments() { //模板中某些函数的调用，我们可以动态添加一些参数。
         return '';
     },
     writeFileStart(e) {
         return e;
     },
-    compileJSStart(content, from) { //开始编译某个js文件之前的处理器，可以加入一些处理，比如typescript的预处理
+    compileJSStart(content) { //开始编译某个js文件之前的处理器，可以加入一些处理，比如typescript的预处理
         return content;
     },
     compileJSEnd(e) { //结束编译
@@ -128,7 +130,7 @@ module.exports = {
     tmplTagProcessor(tag) { //为了tmpl-naked准备的，遇到模板标签如何处理
         return tag;
     },
-    cssNamesProcessor(tmpl, cssNamesMap) { //模板中class名称的处理器
+    cssNamesProcessor(tmpl) { //模板中class名称的处理器
         return tmpl;
     },
     compileTmplCommand(tmpl) {
@@ -149,7 +151,6 @@ module.exports = {
     resolveModuleId(id) { //处理模块id时的处理器
         return id;
     },
-    resolveRequire(reqInfo, context) { //处理rqeuire时的处理器
-        return reqInfo;
+    resolveRequire() { //处理rqeuire时的处理器
     }
 };